refactor(core): Split off ua_securechannel_crypto.c
